Level2/SampleFlask/requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
pathspec==0.8.0
protobuf==3.12.2
regex==2020.5.14
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
WTForms==2.3.1
python-dotenv==0.13.0